config push in sample wallet
