+++ Example/ExampleApp/SceneDelegate.swift
@@ -5,7 +5,6 @@ import WalletConnectSign
 import WalletConnectNetworking
 import WalletConnectRelay
 import WalletConnectPairing
-import WalletConnectPush
 import Combine
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
@@ -14,15 +13,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     private var publishers = [AnyCancellable]()
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-
-        let metadata = AppMetadata(
-            name: "Example Wallet",
-            description: "wallet description",
-            url: "example.wallet",
-            icons: ["https://avatars.githubusercontent.com/u/37784886"])
-
-        Networking.configure(projectId: InputConfig.projectId, socketFactory: DefaultSocketFactory())
-        Pair.configure(metadata: metadata)
 
 
 #if DEBUG
